Use (un)signed short vertices (#90)

+++ internal/graphics/framebuffer.go
@@ -95,8 +95,8 @@ type Matrix interface {
 
 type TextureQuads interface {
 	Len() int
-	Vertex(i int) (x0, y0, x1, y1 float64)
-	Texture(i int) (u0, v0, u1, v1 float64)
+	Vertex(i int) (x0, y0, x1, y1 int)
+	Texture(i int) (u0, v0, u1, v1 int)
 }
 
 func (f *Framebuffer) Fill(c *opengl.Context, r, g, b, a float64) error {
@@ -116,7 +116,7 @@ func (f *Framebuffer) DrawTexture(c *opengl.Context, t *Texture, quads TextureQu
 
 type VertexQuads interface {
 	Len() int
-	Vertex(i int) (x0, y0, x1, y1 float64)
+	Vertex(i int) (x0, y0, x1, y1 int)
 	Color(i int) color.Color
 }
